requirements.txt
# Doctor Image Generator - Python Dependencies

# Core dependencies
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0

# Hugging Face
transformers>=4.30.0
diffusers>=0.25.0
accelerate>=0.20.0
huggingface-hub>=0.19.0

# Image processing
Pillow>=10.0.0
opencv-python>=4.8.0
numpy>=1.24.0
scipy>=1.11.0

# Face detection and processing
insightface>=0.7.3
onnxruntime-gpu>=1.16.0

# API and utilities
requests>=2.31.0
aiohttp>=3.9.0
pyyaml>=6.0
tqdm>=4.66.0

# ComfyUI specific (if running standalone API)
websocket-client>=1.6.0

